plugins/ai_gateway/requirements.txt
anthropic>=0.25.0
openai>=1.0.0
pyyaml>=6.0
